+++ src/main/java/dataFactory/loginUserDataFactory/LoginUserDataFactory.java
@@ -1,0 +1,13 @@
+package dataFactory.loginUserDataFactory;
+
+import dataObject.loginUser.LoginUserDataObject;
+
+public class LoginUserDataFactory {
+
+  public static LoginUserDataObject invalidLoginUserData() {
+    var loginUserData = new LoginUserDataObject();
+    loginUserData.setEmail("invalidEmail");
+    loginUserData.setPassword("invalidPassword");
+    return loginUserData;
+  }
+}
