+++ src/main/java/dataObject/loginUser/LoginUserDataObject.java
@@ -1,0 +1,9 @@
+package dataObject.loginUser;
+
+import lombok.Data;
+
+@Data
+public class LoginUserDataObject {
+  private String email;
+  private String password;
+}
