--- src/test/java/com/selenium/dataObject/loginUser/LoginUserDataObject.java
@@ -1,9 +1,0 @@
-package com.selenium.dataObject.loginUser;
-
-import lombok.Data;
-
-@Data
-public class LoginUserDataObject {
-  private String email;
-  private String password;
-}
