+++ src/test/java/com/selenium/testng/elite/webTest/homePageTest/HomeTest02.java
@@ -2,9 +2,9 @@ package com.selenium.testng.elite.webTest.homePageTest;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import com.selenium.pageObjectModel.webPageObject.checkoutPage.CheckoutPage;
-import com.selenium.pageObjectModel.webPageObject.homePage.HomePage;
-import com.selenium.pageObjectModel.webPageObject.productPage.ProductPage;
+import pageObjectModel.webPageObject.checkoutPage.CheckoutPage;
+import pageObjectModel.webPageObject.homePage.HomePage;
+import pageObjectModel.webPageObject.productPage.ProductPage;
 import com.selenium.testng.elite.BaseTest;
 import com.selenium.testng.elite.enums.EnvironmentType;
 import com.selenium.testng.elite.utils.RandomHelper;
