+++ src/test/java/com/selenium/testng/elite/webTest/registerPageTest/RegisterTest01.java
@@ -2,10 +2,10 @@ package com.selenium.testng.elite.webTest.registerPageTest;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import com.selenium.dataFactory.registerUserDataFactory.RegisterUserDataFactory;
-import com.selenium.pageObjectModel.webPageObject.homePage.HomePage;
-import com.selenium.pageObjectModel.webPageObject.loginPage.LoginPage;
-import com.selenium.pageObjectModel.webPageObject.registrationPage.RegistrationPage;
+import dataFactory.registerUserDataFactory.RegisterUserDataFactory;
+import pageObjectModel.webPageObject.homePage.HomePage;
+import pageObjectModel.webPageObject.loginPage.LoginPage;
+import pageObjectModel.webPageObject.registrationPage.RegistrationPage;
 import com.selenium.testng.elite.BaseTest;
 import java.util.Arrays;
 import org.testng.annotations.Test;
